+++ LoxScript/Core/Scripting/Rules/RuleInvocationContext.cs
@@ -2,16 +2,16 @@
 
 namespace XPT.Core.Scripting.Rules {
     class RuleInvocationContext {
-        private readonly Dictionary<ulong, double> _Variables = new Dictionary<ulong, double>();
+        private readonly Dictionary<long, long> _Variables = new Dictionary<long, long>();
 
-        internal bool TryGetValue(ulong varNameBitString, out double value) {
+        internal bool TryGetValue(long varNameBitString, out long value) {
             if (_Variables.TryGetValue(varNameBitString, out value)) {
                 return true;
             }
             return false;
         }
 
-        internal RuleInvocationContext AddContext(string varName, double value) {
+        internal RuleInvocationContext AddContext(string varName, long value) {
             _Variables[BitString.GetBitStr(varName)] = value;
             return this;
         }
